ref(types): extract AutofixContentProps into a leaf module
types/overrides imported `AutofixContentProps` from the heavy
issueDetails/sidebar/autofixSection view solely for a type annotation in the
override map, tethering that view + its subtree into the frontend type-import
SCC. Move the prop type into a sibling autofixSectionTypes leaf and repoint
importers (overrides + gsApp's aiConfigureSeerQuotaSidebar, which keeps its
runtime AutofixContent import).

Largest type-import SCC: 1657 -> 1638.
