+++ static/gsApp/components/ai/aiConfigureSeerQuotaSidebar.tsx
@@ -7,10 +7,8 @@ import {Placeholder} from 'sentry/components/placeholder';
 import {IconOpen} from 'sentry/icons/iconOpen';
 import {t} from 'sentry/locale';
 import {useOrganization} from 'sentry/utils/useOrganization';
-import {
-  AutofixContent,
-  type AutofixContentProps,
-} from 'sentry/views/issueDetails/sidebar/autofixSection';
+import {AutofixContent} from 'sentry/views/issueDetails/sidebar/autofixSection';
+import type {AutofixContentProps} from 'sentry/views/issueDetails/sidebar/autofixSectionTypes';
 
 import {useSubscription} from 'getsentry/hooks/useSubscription';
 import {hasAccessToSubscriptionOverview} from 'getsentry/utils/billing';
